requirements.txt
allure-pytest==2.8.18
allure-python-commons==2.8.18
Appium-Python-Client==2.7.1
astroid==2.12.12
async-generator==1.10
attrs==22.1.0
certifi==2022.9.24
charset-normalizer==2.1.1
dill==0.3.6
exceptiongroup==1.0.1
execnet==1.9.0
h11==0.14.0
idna==3.4
iniconfig==1.1.1
isort==5.10.1
lazy-object-proxy==1.8.0
mccabe==0.7.0
outcome==1.2.0
packaging==21.3
platformdirs==2.5.4
pluggy==1.0.0
py==1.11.0
pylint==2.15.5
pyparsing==3.0.9
PySocks==1.7.1
pytest==6.2.5
pytest-base-url==2.0.0
pytest-html==3.2.0
pytest-metadata==2.0.4
pytest-selenium==4.0.0
pytest-variables==2.0.0
pytest-xdist==3.0.2
requests==2.28.1
selenium==4.6.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
tenacity==6.3.1
termcolor==2.1.0
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.6
trio==0.22.0
trio-websocket==0.9.2
typing-extensions==4.4.0
urllib3==1.26.12
wrapt==1.14.1
wsproto==1.2.0
